+++ src/pybind/mgr/requirements.txt
@@ -1,17 +1,3 @@
-pytest-cov==2.7.1
--e../../python-common
-kubernetes
-requests-mock
-pyyaml
-prettytable
-pyOpenSSL
-Jinja2
-pyfakefs
+-rrequirements-required.txt
 asyncssh
-asyncmock
-
-jsonpatch
-pecan
-scipy
-cherrypy
-werkzeug
+kubernetes
