+++ src/pybind/mgr/tox.ini
@@ -6,6 +6,7 @@ envlist =
     fix
     flake8
     jinjalint
+    nooptional
 skipsdist = true
 requires = cython
 
@@ -44,6 +45,17 @@ deps =
     -rrequirements.txt
 commands =
     pytest --doctest-modules {posargs:}
+
+[testenv:nooptional]
+setenv =
+    UNITTEST = true
+    PYTHONPATH = $PYTHONPATH:..
+deps =
+    cython
+    -rrequirements-required.txt
+commands =
+    pytest {posargs:cephadm/tests/test_ssh.py}
+
 
 [testenv:mypy]
 setenv =
